refactor(#112): 로그아웃 로직 profileMenu 컴포넌트 내부로 이동
