+++ src/components/Header/ProfileMenu.tsx
@@ -1,9 +1,31 @@
 import Link from 'next/link';
+import { useRouter } from 'next/navigation';
+import { toast } from 'react-toastify';
 import { TransFormSVG } from '@/icons/svg';
+import { getSwitchUserRole, getLogout } from '@/lib/apis/userApi';
 import useSession from '@/lib/useSession';
+import useStore from '@/store';
 
-const ProfileMenu = ({ handleLogout }: { handleLogout: () => void }) => {
+const ProfileMenu = () => {
+  const store = useStore();
   const user = useSession();
+  const router = useRouter();
+
+  const handleLogout = async () => {
+    store.setRequestLoading(true);
+    try {
+      const { status } = await getLogout();
+      if (status === 200) {
+        toast.success('로그아웃 되었습니다!');
+        store.setUserType(null);
+      }
+    } catch (error) {
+      toast.error('로그아웃에 실패하였습니다!');
+    } finally {
+      store.reset();
+      router.push('/');
+    }
+  };
 
   return (
     <ul className="absolute right-0 top-3 inline-flex w-[9.4375rem] select-none flex-col bg-white shadow-vertical">
@@ -21,10 +43,9 @@ const ProfileMenu = ({ handleLogout }: { handleLogout: () => void }) => {
       </li>
 
       <li className="border-t border-solid border-sub-color2 text-main-color">
-        <button className="flex h-full w-full gap-1 py-4 pl-4">
+        <button className="flex h-full w-full cursor-pointer gap-1 py-4 pl-4">
           <TransFormSVG />
-          강사 전환
-          {/* 강사 상태 추후 변경 예정 */}
+          {store.userType === 'user' ? '강사 전환' : '유저 전환'}
         </button>
       </li>
